Organize UTs and integration tests
