+++ ActiveLabelTests/IntegrationTests/BaseIntegrationTestCase.swift
@@ -9,22 +9,7 @@
 import XCTest
 @testable import ActiveLabel
 
-extension ActiveElement: Equatable {}
-
-public func ==(a: ActiveElement, b: ActiveElement) -> Bool {
-    switch (a, b) {
-    case (.mention(let a), .mention(let b)) where a == b: return true
-    case (.hashtag(let a), .hashtag(let b)) where a == b: return true
-    case (.url(let a), .url(let b)) where a == b: return true
-    case (.custom(let a), .custom(let b)) where a == b: return true
-    case (.phone(let a), .phone(let b)) where a == b: return true
-    case (.address(let a), .address(let b)) where a == b: return true
-    case (.date(let a), .date(let b)) where a == b: return true
-    default: return false
-    }
-}
-
-class BaseTestCase: XCTestCase {
+class BaseIntegrationTestCase: XCTestCase {
     let label = ActiveLabel()
     let customEmptyType = ActiveType.custom(pattern: "")
